utils: Move math utils to dedicated file
