+++ tests/unit/external_memory_metal.cpp
@@ -11,11 +11,11 @@
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
 
+#ifdef VK_USE_PLATFORM_METAL_EXT
 #include "utils/vk_layer_utils.h"
 #include "../framework/layer_validation_tests.h"
 #include "../framework/external_memory_sync.h"
 
-#ifdef VK_USE_PLATFORM_METAL_EXT
 // We need these instead of using FindSupportedExternalMemoryHandleTypes because otherwise we'll get
 // VUID-VkPhysicalDeviceExternalImageFormatInfo-handleType-parameter due to sending flags that are not supported due to extensions
 // not being present
