+++ tests/unit/external_memory_sync_positive.cpp
@@ -13,7 +13,7 @@
 
 #include "../framework/layer_validation_tests.h"
 #include "../framework/external_memory_sync.h"
-#include "utils/vk_layer_utils.h"
+#include "utils/math_utils.h"
 #include "generated/enum_flag_bits.h"
 
 class PositiveExternalMemorySync : public ExternalMemorySyncTest {};
