+++ tests/unit/ray_tracing.cpp
@@ -16,7 +16,7 @@
 #include "../framework/ray_tracing_objects.h"
 #include "../framework/descriptor_helper.h"
 #include "../framework/shader_helper.h"
-#include "../layers/utils/vk_layer_utils.h"
+#include "../layers/utils/math_utils.h"
 
 // Compute a binomial coefficient
 template <typename T>
